pyproject.toml
[project]
name = "bparb"
version = "0.1.0"
description = "Backpack Exchange spot-futures arbitrage trading system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.25.0",
    "websockets>=12.0",
    "pynacl>=1.5.0",
    "pydantic>=2.0.0",
    "asyncio-mqtt>=0.13.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0"
]


[tool.ruff]
line-length = 88
target-version = "py312"
